+++ lizmap/server/lizmap_filter.py
@@ -1,59 +1,39 @@
-__copyright__ = 'Copyright 2020, 3Liz'
+__copyright__ = 'Copyright 2021, 3Liz'
 __license__ = 'GPL version 3'
 __email__ = 'info@3liz.org'
 
 import json
 import os
-import traceback
 
 from typing import List
 
-from qgis.core import Qgis, QgsMessageLog
-from qgis.PyQt.QtCore import QByteArray
-from qgis.PyQt.QtXml import QDomDocument
-from qgis.server import QgsServerException, QgsServerFilter, QgsServerInterface
+from qgis.server import QgsServerFilter, QgsServerInterface
 
-
-class LizmapFilterException(QgsServerException):
-
-    def __init__(
-            self, code: str, message: str, locator: str = '', response_code: int = 500, version: str = '1.3.0') -> None:
-        super(QgsServerException, self).__init__(message, response_code)
-        self.code = code
-        self.message = message
-        self.locator = locator
-        self.response_code = response_code
-        self.version = version
-
-    def formatResponse(self) -> (QByteArray, str):
-        doc = QDomDocument()
-        root = doc.createElement('ServiceExceptionReport')
-        root.setAttribute('version', self.version)
-        root.setAttribute('xmlns', 'http://www.opengis.net/ogc')
-        doc.appendChild(root)
-
-        elem = doc.createElement('ServiceException')
-        elem.setAttribute('code', self.code)
-        elem.appendChild(doc.createTextNode(self.message))
-        root.appendChild(elem)
-
-        if self.locator:
-            elem.setAttribute('locator', self.locator)
-
-        return doc.toByteArray(), 'text/xml; charset=utf-8'
+from lizmap.server.exception import LizmapFilterException
+from lizmap.server.logger import Logger
 
 
 class LizmapFilter(QgsServerFilter):
 
-    def __init__(self, server_iface: 'QgsServerInterface') -> None:
-        QgsMessageLog.logMessage('LizmapFilter.init', 'lizmap', Qgis.Info)
+    def __init__(self, server_iface: QgsServerInterface) -> None:
+        Logger.info('LizmapFilter.init')
         super().__init__(server_iface)
 
         self.iface = server_iface
 
     def requestReady(self):
+        logger = Logger()
         # noinspection PyBroadException
         try:
+            # Check first the headers to avoid unnecessary config file reading
+            # Get Lizmap user groups defined in request headers
+            groups = self.getLizmapGroups()
+
+            # If groups is empty, no Lizmap user groups provided by the request
+            # The request can be evaluated by QGIS Server
+            if len(groups) == 0:
+                return
+
             # Get QGIS Project path
             config_path = self.iface.configFilePath()
             if not os.path.exists(config_path):
@@ -65,16 +45,8 @@ class LizmapFilter(QgsServerFilter):
             config_path += '.cfg'
             if not os.path.exists(config_path):
                 # Lizmap config path does not exist
-                QgsMessageLog.logMessage("Lizmap config does not exist", "lizmap", Qgis.Info)
+                logger.info("Lizmap config does not exist")
                 # The request can be evaluated by QGIS Server
-                return
-
-            # Get Lizmap user groups defined in request headers
-            groups = self.getLizmapGroups()
-
-            # If groups is empty, no Lizmap user groups provided by the request
-            # The request can be evaluated by QGIS Server
-            if len(groups) == 0:
                 return
 
             # Get Lizmap config
@@ -84,20 +56,20 @@ class LizmapFilter(QgsServerFilter):
                     cfg = json.loads(cfg_file.read())
                 except Exception:
                     # Lizmap config is not a valid JSON file
-                    QgsMessageLog.logMessage("Lizmap config not well formed", "lizmap", Qgis.Critical)
+                    logger.critical("Lizmap config not well formed")
                     # The request can be evaluated by QGIS Server
                     return
 
             if not cfg:
                 # Lizmap config is empty
-                QgsMessageLog.logMessage("Lizmap config is empty", "lizmap", Qgis.Warning)
+                logger.warning("Lizmap config is empty")
                 # The request can be evaluated by QGIS Server
                 return
 
             # Check Lizmap config options
             if 'options' not in cfg or not cfg['options']:
                 # Lizmap config has no options
-                QgsMessageLog.logMessage("Lizmap config has no options", "lizmap", Qgis.Warning)
+                logger.warning("Lizmap config has no options")
                 # The request can be evaluated by QGIS Server
                 return
 
@@ -105,13 +77,13 @@ class LizmapFilter(QgsServerFilter):
             cfg_options = cfg['options']
             if 'acl' not in cfg_options or not cfg_options['acl']:
                 # No acl defined
-                QgsMessageLog.logMessage("No acl defined in Lizmap config", "lizmap", Qgis.Info)
+                logger.info("No acl defined in Lizmap config")
                 # The request can be evaluated by QGIS Server
                 return
 
             # Get project acl option
             cfg_acl = cfg_options['acl']
-            QgsMessageLog.logMessage("Acl defined in Lizmap config", "lizmap", Qgis.Info)
+            logger.info("Acl defined in Lizmap config")
 
             # If one Lizmap user group provided in request headers is
             # defined in project acl option, the request can be evaluated
@@ -129,10 +101,10 @@ class LizmapFilter(QgsServerFilter):
             # use setServiceException to be sure to stop the request
             handler.setServiceException(exc)
 
-        except Exception:
-            QgsMessageLog.logMessage("Unhandled exception:\n{}".format(traceback.format_exc()), "lizmap", Qgis.Critical)
+        except Exception as e:
+            logger.log_exception(e)
 
-    def getLizmapGroups(self) -> 'List[str]':
+    def getLizmapGroups(self) -> List[str]:
         """ Get Lizmap user groups provided by the request """
         # Defined groups
         groups = []
@@ -140,22 +112,24 @@ class LizmapFilter(QgsServerFilter):
         # Get request handler
         handler = self.iface.requestHandler()
 
+        logger = Logger()
+
         # Get Lizmap User Groups in request headers
         headers = handler.requestHeaders()
         if headers:
-            QgsMessageLog.logMessage("Request headers provided", "lizmap", Qgis.Info)
+            logger.info("Request headers provided")
             # Get Lizmap user groups defined in request headers
             user_groups = headers.get('X-Lizmap-User-Groups')
             if user_groups is not None:
                 groups = [g.strip() for g in user_groups.split(',')]
-                QgsMessageLog.logMessage("Lizmap user groups in request headers", "lizmap", Qgis.Info)
+                logger.info("Lizmap user groups in request headers")
         else:
-            QgsMessageLog.logMessage("No request headers provided", "lizmap", Qgis.Info)
+            logger.info("No request headers provided")
 
         if len(groups) != 0:
             return groups
         else:
-            QgsMessageLog.logMessage("No lizmap user groups in request headers", "lizmap", Qgis.Info)
+            logger.info("No lizmap user groups in request headers")
 
         # Get group in parameters
         params = handler.parameterMap()
@@ -164,6 +138,6 @@ class LizmapFilter(QgsServerFilter):
             user_groups = params.get('LIZMAP_USER_GROUPS')
             if user_groups is not None:
                 groups = [g.strip() for g in user_groups.split(',')]
-                QgsMessageLog.logMessage("Lizmap user groups in parameters", "lizmap", Qgis.Info)
+                logger.info("Lizmap user groups in parameters")
 
         return groups