+++ CHANGELOG.md
@@ -2,7 +2,37 @@
 
 ## Unreleased
 
-## 3.5.2 - 2010-06-23
+* If running QGIS 3.10 or newer, display the file name when using a drag&drop layout with an attachement widget
+* Refactor some code on the server side
+
+## 3.5.7 - 2021-08-31
+
+* Server - Avoid issue about GetFeatureInfo
+* Server - Refactor some code about logging
+
+## 3.5.6 - 2021-08-10
+
+* Server - Avoid registering twice the Lizmap service
+* Server - Add some more info level message in the server
+
+## 3.5.5 - 2021-08-09
+
+* Desktop - Show dialog in front of QGIS Desktop if a dialog is already opened (contribution from @Kanahiro)
+* Server - Improve debug for GetFeatureInfo on the server side
+* Server - Check headers content before reading config file to improve performance
+
+## 3.5.4 - 2021-07-13
+
+* Fix the visibility used on container in QGIS Form with level upper than 1
+* Server - Always log when there is a Python exception
+* Server - Make the plugin compatible Python 3.5
+* Server - Rename variable about conflict
+
+## 3.5.3 - 2021-07-05
+
+* Refactor some imports and class names in the server part
+
+## 3.5.2 - 2021-06-23
 
 * Fix an issue on QGIS Desktop about a Python exception again